requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.4
fakeredis==1.0.4
Flask==1.1.1
flask-redis==0.4.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycryptodomex==3.8.2
pyparsing==2.4.2
pytest==5.1.1
python-dotenv==0.10.3
redis==3.3.8
six==1.12.0
sortedcontainers==2.1.0
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
